deeprxiv-backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
pydantic
requests
PyPDF2
opencv-python
numpy
python-dotenv
pillow
pdf2image
google-genai
chromadb
uuid
PyMuPDF
# Authentication & Security
passlib[bcrypt]
python-jose[cryptography]
authlib
httpx
# Analytics
posthog
# Email
email-validator
# CORS
fastapi-cors
# Rate limiting
slowapi
# Caching
redis
# Session management
fastapi-sessions